Attempted fix for flaky postgres tests
Summary: Logs in the flaky postgres test suggests that the issue happens when you start watching a run, and then immediately fire an event (before it has time to actually set up the listeners). This is an attempt to mitigate that by including an event that the watcher thread uses to signal when it has actually started listening.

Test Plan: BK (several times to try to catch the flakiness)

Reviewers: alangenfeld, sidkmenon

Differential Revision: https://dagster.phacility.com/D6965