prevent start/stop of sensors from messing up tick timestamps
Summary: We were writing tick timestamps when we were turning sensors on/off.  That already doesn't
make sense, but we were also writing it in a different format (i.e. not pendulum.now('UTC')).

This diff makes it so that tick timestamps are only written in one place and on/off operations
just affect status.

Test Plan: added test for sensor on/off behavior

Reviewers: dgibson, sashank

Subscribers: