Reconstruct ExecutionPlan from the ExecutionPlanSnapshot in the run worker if one exists (take 2)
Summary:
When a persisted ExecutionPlanSnapshot is on the run, re-create it from the snapshot rather than rebuilding it from scratch. This is one step on the journey towards being able to have a run worker that doesn't execute any user code and could run in the Dagster cloud, interfacing with step workers in the user cloud.

Requires moving a whole bunch of user code out of ExecutionStep/StepInput/StepOutput and just putting handles on those classes instead.

Test Plan: Integration

Reviewers:

Differential Revision: https://dagster.phacility.com/D6133