[DRAFT] Asset Lineage Rework
Summary:
Imagining a new way to calculate asset lineage that front-loads most of the work. This structure is much more flexible and decipherable (allows you to easily see the assumptions we're making). Dynamic parts of the graph are going to be a bit annoying, but I'm pretty confident there's nothing about them that can't be captured in this model (we'll just have to resolve them based on the asset key later on).

Notably, this allows for transitive lineage (which I had to remove from the 0.11.0 release due to performance/resource scoping issues).

Test Plan:

bk

Reviewers:
prha, sandyryza

Subscribers: