Roll user deployments if Helm values change
Summary:
Resolves https://github.com/dagster-io/dagster/issues/3903.

The user deployment should redeploy if its dependent configmap changes.
However, currently it will not do this since the deployment template
will not have changed even if the configmap changes.

To force the deployment to update, we reason that the deployment should
change if its Helm values have changed.

Test Plan: pytest

Reviewers: catherinewu, johann

Differential Revision: https://dagster.phacility.com/D7125