start storing last_materialization event record data
Summary: It was kind of a miss to store just the materialization data instead of the event record.
This diff starts storing the full event record, to minimize any extra fetching needed.

I think the full migration path for this table should be pretty straightforward, and can be done
pretty painlessly with `0.12.0`.  That should remove all the backcompat logic in this diff.

Test Plan: bk

Reviewers: sandyryza, alangenfeld

Subscribers: