[dagit] Only show GanttChart websocket message if closed
Summary:

Since my change to add debouncing on Websocket status changes, an undesired effect is that the initial connection takes the debounced amount of time to take effect for components that observe the Websocket status.

A resulting bug: when kicking off a new run, the Websocket status message on the Gantt chart briefly flashes while the context state is still "Connecting".

To resolve this, just use the `CLOSED` state to determine whether to show the message. This is more correct than using `OPEN` anyway, since `CONNECTING` and `CLOSING` are still not "closed."

A more thorough solution would probably be to use debouncing only when the Websocket has closed, but otherwise update immediately. This should make things better for now, though.

Test Plan:

Kick off a run, verify that the Gantt chart websocket warning does not appear.

Reviewers:

dgibson, bengotow

Subscribers: