[dynamic] fix multiple collects on one step
Summary: This allows for having a solid `collect` on two different dynamic outputs.

To do this we change `resolvable_map` to key on a set of step keys that must be complete instead of just one.

This makes a small step towards supporting being downstream of multiple multiple outputs, but there is still quite a bit of work to make that happen.

Test Plan: added test

Reviewers:

Subscribers: