[dagit] Left nav experiment cleanup
Summary:

Resolves #3837

We're sticking with the current left nav, at least for now, so there's no reason to keep the experimental one in the codebase. We can always resurrect it later if it makes sense.

This diff also removes the "All pipelines" and "Settings" pages, which are not available through the current Dagit navigation. Again, we can resurrect them if needed.

Test Plan:

Buildkite

`yarn workspace @dagit/core find-dead-code`

Reviewers:

bengotow, sandyryza, prha, sashank

Subscribers: