Incorporate run config into chooseing the executor in host mode run worker flow
Summary:
Instead of just passing in the Executor to execute_run_host_mode, let callsites pass in a function that lets you specify the allowed ExecutorDefinitions (as well as their names). The run worker then uses them to create the executor.

Still missing from this is something to enforce config validation in callsites that will be using this run worker (for example, the instance could define the set of allowed executors and the set of default executors,, and at pipeline snapshot creation time and  plan creation time we could throw if the pipeline isn't configured correctly)

Test Plan: BK

Reviewers: alangenfeld, schrockn

Subscribers: cdecarolis

Differential Revision: https://dagster.phacility.com/D7317