RFC: Allow storages to specify a custom sqlalchemy metadata class
Summary:
The use case here would be to customize the metadata class that is used to define (and migrate) the schema used by all our various storages.

We could use this in a more managed environment where we're confident that we want to update run storage, event storage, schedule storage (and maybe even more tables that we define ourselves outside of the core dagster repo) in lockstep (and disable the 'auto-create on startup' feature).

For now I just ported over RunStorage to see what it felt like. Does this seem like a crazy direction?

Reviewers: alangenfeld, prha, max, sashank, johann

Test Plan: BK