execution context naming revamp, remove resources and intermediate storage from run worker
Summary: This diff attempts a complete system-execution-context naming and structure revamp, such that we distinguish between "orchestration" (coordination of processes to perform actual execution of steps) and "execution" (actual execution of steps). As a side-effect, resources and intermediate storage are no longer initialized in the run process.

Test Plan: A ton of unit tests changes

Reviewers: alangenfeld, dgibson, sandyryza

Subscribers: