[dagit] Break apart Asset materialization query
Summary:
When viewing an asset, a single query is performed to retrieve 200 asset materializations. If this is slow, it blocks the entire page from rendering anything useful.

Instead of this, break it into two queries: one to retrieve the latest materialization (limit: 1) and one to retrieve the latest 200. This way, the latest materialization information can be rendered while the rest is still being retrieved.

Test Plan: View an asset, verify that the details and list/graph render properly.

Reviewers: bengotow, sandyryza, prha

Subscribers: