Fix the partition “Run steps” sidebar which required immediate step list
Summary: The "Run Steps" sidebar shown beside the graphs on the parititon view was broken (I think) by the changes we made to load the data on the page asynchronously. The component keeps selection in a `{stepKey: visible=true/false}` dict and the dict was being initialized with no items and not re-initialized when the items loaded.

I think this could be fixed with a useEffect, but I went ahead and split the value into `all` and `hidden`, so that "hidden" can start as an empty array and new step keys discovered during loading are shown by default, which I think is a bit cleaner.

Test Plan: Interact with pipeline partitions page

Reviewers: @dish

Subscribers: