Remove pyrsitent from dagster
Summary: This is no longer needed as a synchronization primitive. It is used exclusively for the in-memory log case. This is only used for the ephmeral instance case. Theorectically someone could manually launch concurrent runs in different threads and share the event log, However this is pretty fringe and I don't think we should promise that that works.

Test Plan: BK and run a run in dagit

Subscribers: