[dagit] Assets: Only show latest materialization in list
Summary:

When viewing a list of an asset's partitioned materializations, bucket them by partition and only show the most recent one in the table. If there are predecessor materializations (by timestamp), link to a dialog that displays these in the same style of table.

Additionally:

- Put the partition name in the leftmost column, as it is the most key-like value.
- Sort the table by partition name, with partition-less materializations at the bottom.

Test Plan:

View an asset, switch to the list view of materializations. For partitions (or non-partiioned materializations) with predecessors, verify correct rendering of ButtonLink and its dialog.

Reviewers:

sandyryza, prha, bengotow

Subscribers: