download csv in tutorial
Summary:
"When should I use a solid and when should I use a regular Python function?" is a question that came up a couple times last week. I believe the examples in our current tutorial give a misleading answer to this question, which makes it difficult for prospective users to reason about the value of Dagster.

This diff posits that the central reason why it's useful to split code across multiple solids is so that it's easy to re-run some code without re-running everything.

In light of that, it replaces the solid that reads a CSV with a solid that downloads a CSV - it's useful to be able to re-run downstream steps without re-downloading the CSV.

This has some small side benefits as well:
* Removes a tutorial setup step.
* Removes reading from `__file__`, which might throw some readers off

It also includes a change that may be even more controversial - it reads the CSV into a Pandas DataFrame instead of a list of dictionaries.  Even for readers who, say, are heavier PySpark users than Pandas users, I believe that they are more likely to be familiar with `pandas.read_csv` than `csv.DictReader`.  Subjectively, I feel like using `csv.DictReader` is the equivalent of using `urllib` instead of `requests` - it just kind of makes us look like we're not hip to what's going on in this domain.  I can split this out to a separate change if too controversial, but I thought it helps to have the whole picture with both of them together.

Test Plan: manual inspection, bk

Reviewers: owen, max, yuhan, sashank

Subscribers: