Correctly classify errors importing user code
Summary:
https://github.com/dagster-io/dagster/issues/3149

You will encounter the error if you load a repository successfuly, then make a change to a pipeline that makes it impossible to import (so a syntax error, or an invalid import, or even something dumb like dividing by zero at the top of the file), then try to run the pipeline.

Previously, all errors of this form were classified the same, as framework errors, which can be confusing because the error is actually on the user's end.

Test Plan: bk

Reviewers: sandyryza, alangenfeld

Differential Revision: https://dagster.phacility.com/D7501