[dagit] Bucket materializations in asset grid
Summary:

Resolves #3951.

A followup to https://dagster.phacility.com/D7487, which used materialization bucketing on the list view. This just does the same on the grid view.

Test Plan:

View assets that have previous materializations for certain partitions, and assets that do not. Verify that the extra row appears at appropriate times in both grid and list views.

Reviewers:

sandyryza, prha

Subscribers: