[dagit] Show permanent pipeline tags on playground
Summary:

Resolves #3755.

In the Dagit pipeline playground, display tags that are defined in code, e.g. https://docs.dagster.io/concepts/solids-pipelines/pipelines#pipeline-tags).

Test Plan:

Add a tag to a pipeline in the toys repo. Open the playground for that pipeline in Dagit, verify that the tag appears. Open the tag editor, verify that the permanent tag is rendered and that I can add/remove session tags.

Reviewers:

bengotow, prha, sashank

Subscribers: