[dagit] Increase hit area to expand error logs
Summary:

Resolves #4065

Increase the hit area for the "View full message" affordance on the run log viewer. Also make it look a little nicer, center it in the row, and make it a `<button>`.

Test Plan:

View a failed run, find a long error in a row. Verify button rendering and behavior.

Reviewers:

bengotow, sandyryza, prha

Subscribers: