[dagit] Preserve `view` value in Asset breadcrumbs
Summary: When using breadcrumbs on Asset directory pages, the view isn't preserved and it jumps to a flattened (but filtered) list that is not accessible through normal navigational means. I think it would be more correct to preserve the directory view, so that using the breadcrumbs keeps you in the same context.

Test Plan: View assets, switch to directory view. Go into a couple levels, then use breadcrumbs to jump back out. Verify that directory view remains.

Reviewers: prha, bengotow

Differential Revision: https://dagster.phacility.com/D7579