[build_resources improvements 2/n] build_resource
Summary: This diff adds a build_resource method on top of build_resources. Most users will only ever have to use the build_resources API(s) for testing the initialization and fxnality of their resources. For these users, it is annoying to have to boilerplate out the creation of a Resources context object, and then get the resource off of that object. This provides a nice interface for those use cases.

Test Plan: Added unit tests

Reviewers: alangenfeld, sandyryza

Differential Revision: https://dagster.phacility.com/D7601