[dagster-postgres] [easy] Quick fix in listen/notify logic
Summary:
There was an unclear `isinstance` call on the notifications yielded from `await_pg_notifications` as a NoneType check - a better pattern is the `is None` check.

I think the reason this was there in the first place where older versions of the `await_pg_notifications` code would occasionally yield a integer signal.

Test Plan: bk

Reviewers: rexledesma

Differential Revision: https://dagster.phacility.com/D7602