Add sensor cursor (independent from run_key), to skip unnecessary sensor evaluation
Summary: Currently there's no way to reduce the search space of a sensor evaluation function without
requesting a run.  This is because there's no way to pass a run_key to advance a cursor when you
want to skip.  This diff separates out the idempotence mechanism (run_key) with the sensor
evaluation checkpointing (cursor).

Test Plan: bk

Reviewers:

Subscribers: