[dagit] Display partition name tooltip if too long in matrix table
Summary:

Resolves #3457

In some cases, partition names are a bit too long to fit within the allotted space on the run matrix. To alleviate this, let's show a tooltip on title names that are too long to fit in the container, based on some quick post-mount dimension calculations.

Test Plan:

View partitions tab for a pipeline, force partition names to be too long to fit in the matrix header. Hover on text, verify that tooltip appears. Repeat hover for names that aren't too long, verify that there is no tooltip.

Reviewers:

bengotow, prha

Subscribers: