[4/n] Fix inclusive logic when generating partitions
Summary:
Currently, the inclusive flag just removes the last partition
if it is enabled. This looks like incorrect logic.

Instead, we use the inclusive flag to determine if we stop when
we are equal or strictly greater than the end date timestamp.

Test Plan:
fix existing tests

Reviewers:
dgibson, prha, sidkmenon

Subscribers: