[5/n] Fix mapping from execution time to partition
Summary:
When mapping from execution time to partition value,
we disregard the minute/hour values by setting each to 0.

This results in incorrect behavior if we format the partition
value to include those minute/hour values.

Here, we remove the override to 0, so that we defer those values to
the cron schedule minute/hour settings.

Test Plan:
new pytest that failed before this change

Reviewers:
dgibson, prha, sidkmenon

Subscribers: