Scaffold networking and load balancer
Summary:
This introduces a CloudFormation stack that leverages some of AWS's ECS
reference architecture templates:

https://github.com/aws-samples/ecs-refarch-cloudformation

Additional commits will add the Dagster specific resources (an ECS
cluster, a daemon service, a dagit service, an example user repository,
an S3 bucket, and an RDS database).

With this CloudFormation template, users should be able to trivially get
a simple deployment of Dagster running on ECS following our best
practices. Think of it as the ECS equivalent of our Helm charts for K8s.

Test Plan: Create the stack and look at the resources it creates

Reviewers: dgibson, max, johann, rexledesma

Differential Revision: https://dagster.phacility.com/D7665