Add a Fargate cluster
Summary:
For now, everything will be defined in the parent stack. Eventually,
each additional resource will be pulled out into child stacks. To do
this, additional release tooling has to be written to publish stacks to
an S3 bucket so they can be referenced with `templateURL`:

https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-stack.html#aws-properties-stack-properties

Test Plan: Update the stack and check to see that we now have a cluster

Reviewers: dgibson, max, johann, rexledesma

Differential Revision: https://dagster.phacility.com/D7666