Add a dagit service
Summary:
This creates an ECS Service to run Dagit behind our load balancer.
There's significant hardening that needs to go into this before we
release it for actual use.

The particular resources are based in part off of AWS's reference
architecture for a webservice:

https://github.com/aws-samples/ecs-refarch-cloudformation/blob/master/services/website-service/service.yaml

And in part off of what AWS Copilot scaffolds if you stand up a Load
Balanced Service:

https://aws.github.io/copilot-cli/docs/manifest/lb-web-service/

Test Plan:
Update the stack and check to see that we can now view dagit

Reviewers:
dgibson, max, johann, rexledesma

Subscribers: