Re-rache repository locations on WorkspaceProcessContext everytime a request is made
Summary:
Previously, the repository locations from the workspace were cached on the `WorkspaceProcessContext` when initialized. This is because the process context assumes that the workspace only changes when `reload_workspace` is called on the `ProcessContext`. With the new async workspaces, this is not true anymore.

This diff changes the process context so that the locations are re-cached every time a request is made. Alternatively, we can add a thread here that polls for changes on the list of locations on the workspace and updates them proactively, but that might be more complicated than is necessary.

Test Plan: Existing tests

Reviewers: dgibson

Differential Revision: https://dagster.phacility.com/D7683