[Depends on D7670] Allow solid to wrap fxn without context arg
Summary:
A big point of discussion in the solid invocation discussions has been what to do with solids that don't require a context. Ultimately, we realized that a solution might be to permit solids to not have a context arg if it is not being used. This means less magic for a user to define their solid, and be able to invoke it without a context.

This diff makes it possible for the functions that solids wrap to have no context argument. This ideally should only work when no config schema is provided, and there are no required resource keys.

Test Plan: Migrating over conspicuous unit tests.

Reviewers: alangenfeld, sandyryza, schrockn

Differential Revision: https://dagster.phacility.com/D7674