make node result match v1 of results.json
Summary:
Fix for https://github.com/dagster-io/dagster/issues/3815.

The schema of `run_results.json` has changed as of DBT 0.19.0. This breaks the asset materializations in our dbt solids. This diff re-aligns the way we construct materializations such that materializations are yielded again.

Test Plan: Add to `test_dbt_cli_run` to make sure that the correct asset materializations are yielded (this was previously not covered under test).

Reviewers: sandyryza, max

Differential Revision: https://dagster.phacility.com/D7832