[dagster-graphql][client] Fixing `reload_repository_location` and `submit_pipeline_execution`
Summary: This diff addresses [GH issue 4131)(https://github.com/dagster-io/dagster/issues/4131), in which @owen found that `reload_repository_location` was missing cases in the inline fragments in the underlying GraphQL query (and as a result the client would have a runtime error). When fixing this I also discovered a similar missing case for the `submit_pipeline_execution` method.

Also working on something to programmatically ensure that all inline fragment cases are covered for the client queries so we don't experience a regression in the future, but this diff should be enough to solve the 4131 issue.

Test Plan: bk (added new tests for the formerly missing cases for `reloadRepositoryLocation` and `launchPipelineExecution`)

Reviewers: owen, rexledesma

Subscribers: