Pip install twice
Summary:
make dev_install was failing with:

```
ERROR: Cannot install airline-demo==dev, airline-demo[full]==dev, automation==0.0.1 and dagit==dev because these package versions have conflicting dependencies.
The conflict is caused by:
airline-demo[full] dev depends on dagster
airline-demo dev depends on dagster
automation 0.0.1 depends on dagster
dagit dev depends on dagster==dev
```

I think this is related to https://dagster.phacility.com/D7506. My
theory is that until we have dagster et. al. installed as editable
installs, pip is looking to pypi to resolve its dependencies. Because we
switched master away from having a released version and onto always
having "dev" as a version, pip can never find "dagster==dev" on pypi.

Installation works so long as we can get pip past its dependency
resolution step. One option is to revert to using released versions in
our version.py files on master. This is suboptimal becuase it runs
counter to our new release strategy which releases off of branches and
because it creates confusion since local development doesn't actually
depend on these released versions.

Another option is to no longer pin certain projects (dagit,
dagster-graphql, etc.) to a specific version of dagster. This is
suboptimal because we actually do want these pins to exist for when
people install the projects outside of editable mode.

One workaround we found was to pass --use-deprecated=legacy-resolver.
But pip says that they've deprecated this as of 21.0 and threatens to
remove support for it at any point in the future:

https://pip.pypa.io/en/stable/user_guide/#deprecation-timeline-for-2020-resolver-changes

So while it's a valid workaround, it potentially just kicks the problem
down the road until the functionality is suddenly removed.

The workaround I've settled on for now is to do two installations:

1. Install all of our projects as editable without their dependencies
(passing --no-deps).
2. Install again; this time with dependencies.

While perhaps a bit duplicative, this satisfies pip's dependency
resolution because dagster and friends are already install locally so it
never has to reach out to pypi to see if the "dev" version exists.

Test Plan: I created a fresh 3.7.8 virtualenv, ran `pip install --upgrade` pip (to get me up to pip 21.1.1) and then I ran `make dev_install`

Reviewers: alangenfeld, dgibson, max, sashank

Subscribers: