[dagit] Extract AppCache to top level, consolidate TestProvider
Summary:

Move AppCache out of AppProvider so that it can be more flexible about the supplied schema. Also consolidate providers used for tests and storybook, to make it a little easier to set up specs without a ton of requirements.

Test Plan:

Run Dagit to sanity check that everything is working. Jest, lint, ts.

Reviewers:

bengotow

Subscribers: