Include container_image in recon_pipeline_from_origin and recon_repository_from_origin,
Summary: This was presumably an oversight - these origins have container images, they should be included in the resulting ReconstructableRepository/ReconstructablePipeline.

The result of this is that we can use the origin in various executors rather than relying on DAGSTER_CURRENT_IMAGE being set in the run launcher.

Test Plan: Integration

Differential Revision: https://dagster.phacility.com/D7894