Add loading status to GrapheneRepositoryLocation / repository location list.
Summary:
This adds the graphql and dagit UI components to allow dagit to indicate that locations are in the process of loading, but aren't ready yet. We could use this in cases where reloading a repository location takes a really long time and we don't want to block Dagit on it.

This gets a little bit tricky, b/c you can be loading/updating while still serving the soon-to-be-replaced location or error to queries that are coming in. So the existing types (location and error) can a 'loadStatus' - and then there's a separate RepositoryLocationLoading type that can be used during the *initial* load of a new location, when there's no repository location or error yet to display. In the former case we say "Updating.." and in the latter we say "Loading..."

Open to feedback and alternatives on this.

Test Plan: Test reloading a repository location and workspace, behavior unchanged
Test in a branch that actually uses the UPDATING state, locations transition into Loading... state when being newly created and Updating... after updating a location.

Differential Revision: https://dagster.phacility.com/D7931

Reviewers: sashank, dish