[schedule/sensor testing 1/n] [RFC] build_schedule_context
Summary: This diff provides a convenient fxn for constructing a schedule context. Right now, it is just a function, but there is an argument for making it a context manager that defaults the instance to `instance.get()`.

Test Plan: Convert unit test decorators to use build_schedule_context instead of directly constructing ScheduleExecutionContext

Reviewers: sandyryza, alangenfeld, prha

Differential Revision: https://dagster.phacility.com/D7942