Fix inclusion of partitions mapped from execution time
Summary:
Depends on D7654.

A previous attempt was at https://dagster.phacility.com/D7656.

Inclusion of executed partitions from a schedule tick depends on the
partition offset. However, we naively just popped off the last partition
if any offset existed. If the offset is greater than 1, this logic
would not make any sense.

One approach to fix this is to apply the execution_time_to_partition_fn
on the end time of the schedule tick iterator. This way, any partition
in the range between the offset and the end time would not be included.

Test Plan: Add test cases that would have failed before, fix old tests

Reviewers: dgibson, prha

Differential Revision: https://dagster.phacility.com/D7981