Allow users to launch a backfill of the last pipeline partition set that generated an asset #3874
Summary: This diff adds a "Launch Backfill" button to the "Last materialization from" line of the asset details header, making it easy to launch a backfill from the asset page.

There's a bit of a sublte distinction here - while assets have partitions they do not conform to a partition set and are not bound to a specific pipeline. You can't backfill an asset, but you can backfill (one of the) partition sets which generates an asset. For this UI, if the last materialization was from a partitioned pipeline run we pull the context from that run and offer to backfill more runs of that pipeline + partition set. To make this a bit more clear I put the "Launch Backfill" button next to the pipeline reference instead of next to the list of partitions. If you do have two pipelines generating an asset, the launch backfill button will vary in behavior depending on which materialized the asset last. (eek)

Test Plan: Run tests, give it a spin!

Reviewers: @dish, @sandyryza

Subscribers: