[dagit] Make timestamp column resizable in log viewer
Summary:

Resolves #4193

The timestamp column in the run log viewer is currently not resizable, and the timestamp text is right-aligned. This means that the scrollbar sometimes sits on top of the timestamp text, which is irritating.

Make the column resizable, increase its default width, and make the timestamp text left-aligned. This way the default should keep the timestamp in view while allowing the user to customize as desired.

Test Plan:

View run logs, verify that timestamp column is resizable and left-aligned.

Reviewers:

bengotow, prha

Subscribers: