Replace WorkspaceSnapshot with a dict of WorkspaceLocationEntry
Summary: The number of dicts in WorkspaceSnapshot was getting out of hand (and i want to add timestamp and display metadata so that we can display information about locations that are still loading or failed to load). Refactor WorkspaceSnapshot to be a dict of WorkspaceLocatioEntry instead. Next diff will augment the graphql to expose the new type in the status page, for now nothing user-visible should change.

Test Plan: BK, view status page in dagit

Reviewers: sashank, alangenfeld

Differential Revision: https://dagster.phacility.com/D8050