New GraphQL connection for repository locations that includes display metadata and a timestamp
Summary:
Rather than using a gross union of location+error+loading, keep a consistent type that represents an entry in the workspace. This gives us a good place to attach things like timestamp, display metadata, and load status that stays the same whether or not hte locaito nloaded or not.

Test Plan: BK

Reviewers: alangenfeld, sashank

CC: dish