Add a way to run the python graphql client against the existing graphql_context fixtures (using the real graphql schema)
Summary: this adds a test that https://dagster.phacility.com/D8070 would have broken, and then fixes the underlying query so that it's no longer broken. The unit tests we had before are useful for testing business logic, but not that the query still works against the actual schema.

Test Plan: BK

Reviewers: alangenfeld, sidkmenon, rexledesma

Differential Revision: https://dagster.phacility.com/D8080