Tiny fix to event log links from runs page
Summary:

Noticed that if you went to the runs page and hovered over the "SUCCESS" indicator, it would show you some stats about the run, which in turn had links. These links did not have the proper type names, so if you clicked them, the filter would be empty. This diff corrects the event names.

Test Plan:
bk

Reviewers:
dish, bengotow

Subscribers: