Reorg implementation of config eval on solid invocation
Summary: This diff reorgs the implementation of config evaluation on solid invocation. Essentially, the Shape config type cannot accept a None value ever, and the underlying config type to the solid definition may not permit a dict. The solution is to always wrap in a shape, permitting us to place a dict as default arg
I'm not sure if this is actuall an improvement, but it's at least more clear about what's going on.

Test Plan: unit tests

Reviewers: sandyryza, alangenfeld

Subscribers: