[dagit] Crag: Populate empty config if only one preset
Summary: For Crag flagged users, in cases where there is only one preset for a given job, automatically populate the Playground config editor with that preset yaml and do not render the preset picker. Similarly, if there is only one partition set for the mode, do not render the partition set picker.

Test Plan:
Clear localStorage.

With flag enabled:

- View a pipeline:mode tuple Playground page with only one preset. Verify that the editor is populated with the preset config yaml.
- Create a new tab on the Playground, verify that config is prepopulated.
- View a pipeline:mode tuple Playground page with only one partition set. Verify that the partition set picker is not rendered.

With flag disabled, verify that the above scenarios match the current behavior/rendering in master.

Reviewers: bengotow, prha, sandyryza

Differential Revision: https://dagster.phacility.com/D8643