rename SensorExecutionContext to SensorContext
Summary:
In the recent orchestration / execution context split, the word `execution` has taken on more meaning. While this has not been codified anywhere yet, the current thinking is that an `ExecutionContext` is a context involved in the core computation of the plan; having access to user code, resources, etc.

The SensorExecutionContext does not fit this bill, since it is used to evaluate config, which is still technically during the "orchestration" phase (constructing the plan, routing the steps of the plan, evaluation of config, etc).

One downside is that SensorContext is a bit of a vague name. Perhaps we could still include another word to more accurately reflect its purpose; SensorTickContext is one idea.

Test Plan: Unit and test docs build

Differential Revision: https://dagster.phacility.com/D8362