[dagit] Crag: Don't show partition set selector if only one
Summary:

If a pipeline has a single partition set and the user is in the crag flag, don't show the switcher on the Partitions tab.

Test Plan:

View pipeline with single partition set. Verify that switcher is visible when the flag is off, not rendered when the flag is on.

Reviewers:

bengotow, prha, sandyryza

Subscribers: